token added

+++ Controller/Doctor.js
@@ -3,9 +3,14 @@ const mongoose = require('mongoose');
 
 const {Doctors} = require("../Model/Doctor");
 
+const {getAuth,setAuth} = require('../Auth/Auth');
+
 
 exports.addDoctor = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
+
         req.body.ImageID = req.file.id; req.body.ImageFileName = req.file.filename; req.body.ImageOriginalName = req.file.originalname;
         console.log(req.body);
         let data = new Doctors(req.body);
@@ -16,9 +21,11 @@ exports.addDoctor = async(req, res,next) => {
         return res.json({text:` Error Found `,error,status:400});
     }
 }
-
+//token : eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVmZTJlMzhiNWVlZTMxNWU4Yzc3NmQ2MyIsImlhdCI6MTYwODc0NDUzMX0.T8Ck96I4qaFnwwwDm_OVJ6Vv2i29ozG5PkssfrWH1t0
 exports.getDoctors = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
         let Doctor = await Doctors.find();
         if(!Doctor) return res.json({text:`No Doctors Added`,status:400});
         return res.send(Doctor);
@@ -31,6 +38,9 @@ exports.getDoctors = async(req, res,next) => {
 
 exports.getDoctor = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
+        
         let Doctor = await Doctors.findById(req.params.id);
         if(!Doctor) return res.json({text:`Doctor Not Found`,status:400});
         return res.send(Doctor);
@@ -43,9 +53,12 @@ exports.getDoctor = async(req, res,next) => {
 
 exports.deleteDoctor = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
+        
         let Doctor = await Doctors.findByIdAndDelete(req.params.id);
         if(!Doctor) return res.json({text:`Doctor Not Found`,status:400});
-        return res.send(Doctor);
+        return res.json({text:"Team Member Deleted",status:200});
 
     } catch (error) {
         return res.json({text:`Error Found`,error,status:400});
@@ -58,6 +71,8 @@ exports.updateDoctor = async(req, res,next) => {
         if(req.file.id != ""){
             req.body.ImageID = req.file.id; req.body.ImageFileName = req.file.filename; req.body.ImageOriginalName = req.file.originalname;
         }
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
         
         let doctor = await Doctors.findById(req.params.id);
         if(!doctor) return res.json({text:`News Not Found`,status:400});
