+++ Controller/News.js
@@ -2,11 +2,15 @@ const express = require('express');
 const mongoose = require('mongoose');
 
 const {News,validateData} = require("../Model/News");
-// const {getAuth,setAuth} = require('../../Auth/auth');
+const {getAuth,setAuth} = require('../Auth/Auth');
+const { json } = require('express');
 
 //get all bank with info
 exports.getNews = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
+
         let news = await News.find();
         if(!news) return res.json({text:`No News Added`,status:400});
         return res.send(news);
@@ -19,6 +23,9 @@ exports.getNews = async(req, res,next) => {
 
 exports.getNew = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
+
         let single_news = await News.findById(req.params.id);
         if(!single_news) return res.json({text:`News Not Found`,status:400});
         return res.send(single_news);
@@ -31,6 +38,9 @@ exports.getNew = async(req, res,next) => {
 
 exports.addNews = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
+
         req.body.ImageID = req.file.id; req.body.ImageFileName = req.file.filename; req.body.ImageOriginalName = req.file.originalname;
         // console.log(req.body,req.file);
         // const {error} = validateData(req.body);
@@ -48,6 +58,9 @@ exports.addNews = async(req, res,next) => {
 
 exports.updateNews = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
+
         if(req.file.id != ""){
             req.body.ImageID = req.file.id; req.body.ImageFileName = req.file.filename; req.body.ImageOriginalName = req.file.originalname;
         }
@@ -66,8 +79,11 @@ exports.updateNews = async(req, res,next) => {
 
 exports.deleteNews = async(req, res,next) => {
     try {
+        let auth = getAuth(req.headers.token);
+        if (!auth.id) return res.json({text:`Token Denied`,status:403});
+
         let news = await News.findByIdAndDelete(req.params.id);
-        res.send("News Deleted")
+        return res.json({text:"News Deleted",status:200})
     } catch (error) {
         return res.json({text:` Error Found `,error,status:400});
     }
